requirements.txt
psycopg==3.1.9
psycopg-binary==3.1.9
python-environ==0.4.54
tabulate==0.9.0